backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
qdrant-client==1.12.0
openai==1.52.0
asyncpg==0.30.0
python-dotenv==1.0.1
tiktoken==0.8.0
python-frontmatter==1.1.0
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
